pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
combine_as_imports = "true"
include_trailing_comma = "true"
known_first_party = ["src"]
line_length = 120
multi_line_output = 3

[tool.poetry]
authors = ["Dominik <*******@gmail.com>"]
description = ""
name = "crm-ddd-poc"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
attrs = "^24.2.0"
email-validator = "^2.2.0"
faker = "^28.4.1"
fastapi = {extras = ["standard"], version = "^0.112.2"}
firebase-admin = "^6.5.0"
phonenumbers = "^8.13.43"
pydantic = "^2.8.2"
python = "^3.12"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.35"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.9"
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pre-commit = "^3.8.0"
pylint = "^3.3.0"
pylint-per-file-ignores = "^1.3.2"
pytest = "^8.3.3"
ruff = "^0.6.6"
toml-sort = "^0.23.1"

[tool.pylint.BASIC]
good-names = "id,i,j,k,db,r"

[tool.pylint.DESIGN]
max-args = 12
max-attributes = 12
min-public-methods = 1

[tool.pylint.FORMAT]
max-line-length = 120

[tool.pylint.MASTER]
init-hook = "import sys; sys.path.append('./src')"
load-plugins = [
  "pylint_per_file_ignores"
]

[tool.pylint."MESSAGES CONTROL"]
allowed-redefined-builtins = ['id']
disable = "missing-docstring, logging-fstring-interpolation, too-few-public-methods, undefined-variable, too-many-positional-arguments"

[tool.pylint.SIMILARITIES]
ignore-comments = "yes"
ignore-docstrings = "yes"
ignore-imports = "yes"
min-similarity-lines = 6

[tool.pylint-per-file-ignores]
"/containers/" = "too-many-instance-attributes"
"customer_status.py" = "protected-access"
"db.py" = "not-callable"
"models.py" = "unsubscriptable-object"

[tool.pytest.ini_options]
markers = [
  "integration"
]
